xe: jit: dsl: move {global_}tensor_t to tensor.hpp
